Revert "Staging to Master. Version 1.10.0 (#4327)"
This reverts commit 7ef004d78ba533a541f7daa5a69606b1e7f2687a.

+++ ironfish/src/memPool/recentlyEvictedCache.ts
@@ -1,7 +1,6 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import Decimal from 'decimal.js'
 import { Logger } from '../logger'
 import { MetricsMonitor } from '../metrics'
 import { TransactionHash } from '../primitives/transaction'
@@ -9,7 +8,7 @@ import { PriorityQueue } from './priorityQueue'
 
 interface EvictionQueueEntry {
   hash: TransactionHash
-  feeRate: Decimal
+  feeRate: bigint
 }
 
 interface RemoveAtSequenceQueueEntry {
@@ -150,7 +149,7 @@ export class RecentlyEvictedCache {
    */
   add(
     transactionHash: TransactionHash,
-    feeRate: Decimal,
+    feeRate: bigint,
     currentBlockSequence: number,
     maxAge: number,
   ): boolean {
